+++ Census Bureau Python/README.md
@@ -1,0 +1,2 @@
+# Power-BI-CensusBureau
+Ongoing project to utilize PySpark within Fabric to ingest Census data, starting with most recent ACS 5-Year Estimate U.S. Census Bureau data across multiple geographies (state - zip code - census tract), through Fabric utilizing Medallion architecture.
