--- README.md
@@ -1,3 +1,0 @@
-# Power-BI-CensusBureau
-
-Project has not been advanced in some time. Tinkering with APIs, for the most part, to get a feel for how best to ingest. I've seen some lovely work in R with API ingestion that I am to replicate with native PBI capability.
